Add source for jdk_1.5.0

+++ src/javax/swing/text/AsyncBoxView.java
@@ -1,7 +1,7 @@
 /*
- * @(#)AsyncBoxView.java	1.14 03/01/23
+ * @(#)AsyncBoxView.java	1.16 03/12/19
  *
- * Copyright 2003 Sun Microsystems, Inc. All rights reserved.
+ * Copyright 2004 Sun Microsystems, Inc. All rights reserved.
  * SUN PROPRIETARY/CONFIDENTIAL. Use is subject to license terms.
  */
 package javax.swing.text;
@@ -27,7 +27,7 @@ import javax.swing.event.DocumentEvent;
  * so that the model is stable while being accessed.
  *
  * @author  Timothy Prinzing
- * @version 1.14 01/23/03
+ * @version 1.16 12/19/03
  * @since   1.3
  */
 public class AsyncBoxView extends View {
@@ -663,7 +663,7 @@ public class AsyncBoxView extends View {
 	if (axis == this.axis) {
 	    return getPreferredSpan(axis);
 	}
-	return Short.MAX_VALUE;
+	return Integer.MAX_VALUE;
     }
